+++ NEWS.md
@@ -1,3 +1,10 @@
+# 12.0.0 Unreleased
+
+* Only requires `userId` for open loans (CIRC-136)
+  * Although it is not possible to create already closed loans
+* Provides `circulation` 4.0 interface (CIRC-136)
+* Requires `loan-storage` interface 5.0 (CIRC-136)
+
 ## 11.0.1 2018-09-03
 
 * Fix module permissions for request collection endpoint (CIRC-134)